add Makefile for build
