+++ README.md
@@ -1,2 +1,10 @@
-# channel-service
+# Channel-Service
 post message, upload file/photo/video
+
+
+## Build
+godep restore
+
+## Run
+./channel-service &
+
